feat(componenthome) : Création de composants pour home
