+++ src/views/Reportages.vue
@@ -1,7 +1,12 @@
 <script setup lang="ts">
+import HomePlaylist from '@/components/HomePlaylist.vue';
+import HomeSlider from '@/components/HomeSlider.vue';
+
 
 </script>
 
 <template>
     <h1>Reportages</h1>
+    <HomeSlider />
+    <HomePlaylist />
 </template>
